requirements.txt
# Uses Python 3.10.10

# Base libraries
matplotlib==3.6.2
numpy==1.23.2
scikit-learn==1.2.0
pandas
scipy
seaborn
tqdm==4.64.0

pydash==7.0.6
python-dotenv==1.0.0

# Pretty logs
rich==13.4.1

# Code style and type-checking
black==22.3.0
flake8==4.0.1
isort==5.12.0
pytest==7.1.2
pylint==2.13
mypy==1.2
pre-commit==3.2.2 # Run `pre-commit install` to set-up pre-commit hooks

git+https://github.com/cooper-org/cooper.git@e23aa46ef8ea2ff0c7b169cbadb5877eab482416

# Pytorch
# --extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0

# --extra-index-url https://download.pytorch.org/whl/cu118
torchvision

# Experiment tracking and reproducibility
ml-collections==0.1.1
wandb==0.16.2

# Submitting SLURM jobs
submitit==1.4.5
